+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## v0.2.4
 * update all node pickers to request 25 per page
 * fix `nodePickerField` double scrollbar and overflow issues
+* clear `nodePickerField` search string on select
 
 
 ## v0.2.3